fix(build): enable strict mode compilation with -Werror
Add -isystem for third-party library includes to suppress warnings from
headers we don't control (LVGL, libhv, spdlog, TinyGL, wpa_supplicant).
This allows `make strict` to catch real issues in project code while
ignoring third-party header warnings.

Fix ~244 conversion warnings across 37 files:
- Sign conversion (int ↔ uint32_t/size_t) for array indices and APIs
- 64-to-32 bit narrowing (chrono, vector::size())
- Float precision (M_PI double→float)
- Variable shadowing (renamed inner scope variables)
- Unused private fields (removed from ui_panel_advanced.h)
- Missing #include <map> in usb_backend_mock.h

All casts use explicit static_cast<> for clarity and type safety.
The build now passes `make strict` with zero warnings/errors.

+++ include/moonraker_request.h
@@ -67,7 +67,7 @@ struct PendingRequest {
     uint32_t get_elapsed_ms() const {
         auto now = std::chrono::steady_clock::now();
         auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(now - timestamp);
-        return elapsed.count();
+        return static_cast<uint32_t>(elapsed.count());
     }
 };
