+++ src/ui_text.cpp
@@ -86,7 +86,7 @@ static void apply_semantic_style(lv_obj_t* label, const char* font_const_name,
     // Apply color
     const char* color_str = lv_xml_get_const(NULL, color_const_name);
     if (color_str && color_str[0] == '#') {
-        uint32_t hex = strtoul(color_str + 1, NULL, 16);
+        uint32_t hex = static_cast<uint32_t>(strtoul(color_str + 1, NULL, 16));
         lv_color_t color = lv_color_hex(hex);
         lv_obj_set_style_text_color(label, color, 0);
     } else {
